modal fully functional

+++ app/Http/Controllers/HomeController.php
@@ -47,6 +47,7 @@ class HomeController extends Controller
     public function create()
     {
         //
+        return "create";
     }
 
     /**
@@ -57,7 +58,20 @@ class HomeController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        $this->validate($request,[
+            'StudentName' =>'required',
+            'Student_id' =>'required'
+        ]);
+        $user = new AppUser;
+        if (!isset($request->status))
+            $user->status = 'waiting';
+        else 
+            $user->status = $request->status;
+        
+        $user->username = $request->StudentName;
+        $user->student_id = $request->Student_id;
+        $user->save();
+        return redirect('/');
     }
 
     /**
@@ -68,7 +82,7 @@ class HomeController extends Controller
      */
     public function show($id)
     {
-        //
+        return "show";
     }
 
     /**
@@ -99,7 +113,6 @@ class HomeController extends Controller
      */
     public function update(Request $request, $id)
     {
-        
         $this->validate($request,[
             'StudentName' =>'required',
             'Student_id' =>'required'
@@ -119,8 +132,15 @@ class HomeController extends Controller
      */
     public function destroy($id)
     {
-        $user = AppUser::find($id);
-        $user->delete();
-        return redirect('/');
+        if(isset($id))
+        {
+            $user = AppUser::find($id);
+            $user->delete();
+            return redirect('/');
+        }
+        else
+        {
+            return redirect('/')->with('message','success');
+        }
     }
 }
